+++ src/julia.c
@@ -3,7 +3,8 @@
 #include <R.h>
 #include <Rinternals.h>
 
-SEXP julia_(SEXP cx_, SEXP cy_, SEXP movex_, SEXP movey_, SEXP zoom_, SEXP size_, SEXP max_iter_) {
+SEXP julia_(SEXP cx_, SEXP cy_, SEXP movex_, SEXP movey_, SEXP zoom_, SEXP size_, SEXP max_iter_,
+            SEXP gamma_, SEXP equalize_) {
 
   int size  = asInteger(size_);
 
@@ -17,13 +18,13 @@ SEXP julia_(SEXP cx_, SEXP cy_, SEXP movex_, SEXP movey_, SEXP zoom_, SEXP size_
 
   int max_iter = asInteger(max_iter_);
 
-  SEXP rarray;
-  PROTECT(rarray = allocVector(RAWSXP, size * size));
-  uint8_t *raw_ptr = RAW(rarray);
-
   double idenom = 1.5/(0.5 * zoom * size);
 
   int max = 1;
+
+  int *iters;
+  iters = calloc(size * size, sizeof(int));
+  int *iter_ptr = iters;
 
   for (int x=0; x < size; x++) {
     for (int y = 0; y < size; y++) {
@@ -43,18 +44,75 @@ SEXP julia_(SEXP cx_, SEXP cy_, SEXP movex_, SEXP movey_, SEXP zoom_, SEXP size_
         max = iter;
       }
 
-      *raw_ptr++ = iter;
+      *iter_ptr++ = iter;
     }
   }
 
-  // Normalise
-  raw_ptr = RAW(rarray);
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Normalise into range [0-255] and cast as byte
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  SEXP rarray;
+  PROTECT(rarray = allocVector(RAWSXP, size * size));
+  uint8_t *raw_ptr = RAW(rarray);
+
+  iter_ptr = iters;
+
   for (int i = 0; i < size * size; i++) {
-    *raw_ptr = (uint8_t)(*raw_ptr * 255/max);
-    raw_ptr++;
+    *raw_ptr++ = (uint8_t)round(*iter_ptr++/(double)max * 255);
+  }
+
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Histogram equalization
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  int equalize = asLogical(equalize_);
+  if (equalize) {
+    uint32_t *hist;
+    hist = (uint32_t *)calloc(256, sizeof(uint32_t));
+    double cumsum[256];
+    raw_ptr = RAW(rarray);
+
+    // Count the pixels at each level
+    for (int i = 0; i < size * size; i++) {
+      hist[raw_ptr[i]]++;
+    }
+
+    // Calculate the cumulative sum of pixels
+    cumsum[0] = hist[0];
+    for (int i = 1; i < 256; i++) {
+      cumsum[i] = cumsum[i-1] + hist[i];
+    }
+
+    // Rescale the cumulative sum by the total pixels and scale by the
+    // max level i.e. 255
+    uint8_t remap[256];
+    for (int i = 0; i < 256; i++) {
+      remap[i] = (uint8_t)floor(cumsum[i] / (size * size) * 255);
+    }
+
+    // Remap pixel values to the equalized bins in 'remap'
+    raw_ptr = RAW(rarray);
+    for (int i = 0; i < size * size; i++) {
+      *raw_ptr = remap[*raw_ptr];
+      raw_ptr++;
+    }
+    free(hist);
   }
 
 
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  // Apply gamma correction
+  //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+  double gamma = asReal(gamma_);
+  if (gamma != 1) {
+    raw_ptr = RAW(rarray);
+    for (int i = 0; i < size * size; i++) {
+      *raw_ptr = (uint8_t)round(pow(*raw_ptr/255.0, gamma) * 255);
+      raw_ptr++;
+    }
+  }
+
+
+  free(iters);
 
   UNPROTECT(1);
   return rarray;