+++ NEWS.md
@@ -1,3 +1,9 @@
+# frac 0.1.1  2020-07-19
+
+* Fixed a bug where iterations above 255 were not being handled correctly.
+* Added options for gamma correction (`gamma`) and histogram equalisation
+  (`equalize`)
+
 # frac 0.1.0  2020-07-16
 
 * Initial release